requirements.txt
Django==4.0.4
djangorestframework==3.14.0
django-filter==22.1
django-redis==5.2.0
psycopg2==2.9.3
redis==4.3.4
Jinja2==3.0.3
